+++ lua/me/keymaps.lua
@@ -8,6 +8,20 @@ vim.keymap.set("n", "<leader>sv", "<C-w>v", { desc = "Split window vertically" }
 vim.keymap.set("n", "<leader>sh", "<C-w>s", { desc = "Split window horizontally" })
 vim.keymap.set("n", "<leader>se", "<C-w>=", { desc = "Make splits equal size" })
 vim.keymap.set("n", "<leader>sx", "<cmd>close<CR>", { desc = "Close current split" })
+vim.keymap.set(
+	"n",
+	"H",
+	":vertical resize -2<CR>",
+	{ desc = "Decrease split horizontal size", noremap = true, silent = true }
+)
+vim.keymap.set(
+	"n",
+	"L",
+	":vertical resize +2<CR>",
+	{ desc = "Increase split horizontal size", noremap = true, silent = true }
+)
+vim.keymap.set("n", "J", ":resize +2<CR>", { desc = "Decrease split horizontal size", noremap = true, silent = true })
+vim.keymap.set("n", "K", ":resize -1<CR>", { desc = "Increase split horizontal size", noremap = true, silent = true })
 
 ---------------------
 -- Tabs management --